+++ .gitignore
@@ -2,4 +2,5 @@
 /coverage
 /node_modules
 /out
+/test-reports
 svn-scm*.vsix
