+++ package.json
@@ -56,6 +56,7 @@
     "@semantic-release/changelog": "^3.0.2",
     "@semantic-release/git": "^7.0.8",
     "@types/glob": "^7.1.1",
+    "@types/istanbul": "^0.4.30",
     "@types/mocha": "^5.2.6",
     "@types/node": "^11.11.3",
     "@types/semver": "^5.5.0",
@@ -69,6 +70,7 @@
     "istanbul": "^0.4.5",
     "jschardet": "^1.6.0",
     "mocha": "^6.0.2",
+    "mocha-junit-reporter": "^1.18.0",
     "prettier": "^1.16.4",
     "prettylint": "^1.0.0",
     "remap-istanbul": "^0.13.0",
